pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contentbank"
version = "0.1.0"
description = "ContentBank — content-addressable persistent storage for TinyLibrary"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"

dependencies = [
    # API server
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    # Data validation
    "pydantic>=2.7",
    # RDF / SHACL
    "rdflib>=6.3",
    "pyshacl>=0.25",
    # Database
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    # IPFS
    "ipfshttpclient>=0.7",
    # Auth / crypto
    "python-jose[cryptography]>=3.3",
    "cryptography>=42",
    # Config
    "pydantic-settings>=2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",        # async test client for FastAPI
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
contentbank = "contentbank.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
